+++ engine/access.go
@@ -1,0 +1,110 @@
+package engine
+
+import (
+	"fmt"
+	"net"
+	"net/http"
+	"os"
+	"strconv"
+	"strings"
+	"time"
+
+	log "github.com/sirupsen/logrus"
+)
+
+func EmptyCommonLogFormatLine() string {
+	// goaccess did not accept this as an "empty" line
+	//timestamp := strings.Replace(time.Now().Format("02/Jan/2006 15:04:05 -0700"), " ", ":", 1)
+	//return "- - - [" + timestamp + "] \"- - -\" - -\n"
+	return "\n"
+}
+
+func EmptyCombinedLogFormatLine() string {
+	// goaccess did not accept this as an "empty" line
+	//timestamp := strings.Replace(time.Now().Format("02/Jan/2006 15:04:05 -0700"), " ", ":", 1)
+	//return "- - - [" + timestamp + "] \"- - -\" - - \"-\" \"-\"\n"
+	return "\n"
+}
+
+// CommonLogFormat returns a line with the data that is available at the start
+// of a request handler. The log line is in NCSA format, the same log format
+// used by Apache. Fields where data is not available are indicated by a "-".
+// See also: https://en.wikipedia.org/wiki/Common_Log_Format
+func (ac *Config) CommonLogFormat(req *http.Request, statusCode, byteSize int) string {
+	username := "-"
+	if ac.perm != nil {
+		username = ac.perm.UserState().Username(req)
+	}
+	host, _, err := net.SplitHostPort(req.RemoteAddr)
+	ip := host
+	if err != nil {
+		ip = req.RemoteAddr
+	}
+	statusCodeString := "-"
+	if statusCode > 0 {
+		statusCodeString = strconv.Itoa(statusCode)
+	}
+	byteSizeString := "0"
+	if byteSize > 0 {
+		byteSizeString = strconv.Itoa(byteSize)
+	}
+	timestamp := strings.Replace(time.Now().Format("02/Jan/2006 15:04:05 -0700"), " ", ":", 1)
+	return fmt.Sprintf("%s - %s [%s] \"%s %s %s\" %s %s", ip, username, timestamp, req.Method, req.RequestURI, req.Proto, statusCodeString, byteSizeString)
+}
+
+// CombinedLogFormat returns a line with the data that is available at the start
+// of a request handler. The log line is in CLF, similar to the Common log format,
+// but with two extra fields.
+// See also: https://httpd.apache.org/docs/1.3/logs.html#combined
+func (ac *Config) CombinedLogFormat(req *http.Request, statusCode, byteSize int) string {
+	username := "-"
+	if ac.perm != nil {
+		username = ac.perm.UserState().Username(req)
+	}
+	host, _, err := net.SplitHostPort(req.RemoteAddr)
+	ip := host
+	if err != nil {
+		ip = req.RemoteAddr
+	}
+	statusCodeString := "-"
+	if statusCode > 0 {
+		statusCodeString = strconv.Itoa(statusCode)
+	}
+	byteSizeString := "0"
+	if byteSize > 0 {
+		byteSizeString = strconv.Itoa(byteSize)
+	}
+	timestamp := strings.Replace(time.Now().Format("02/Jan/2006 15:04:05 -0700"), " ", ":", 1)
+	referer := req.Header.Get("Referer")
+	userAgent := req.Header.Get("User-Agent")
+	return fmt.Sprintf("%s - %s [%s] \"%s %s %s\" %s %s \"%s\" \"%s\"", ip, username, timestamp, req.Method, req.RequestURI, req.Proto, statusCodeString, byteSizeString, referer, userAgent)
+}
+
+func (ac *Config) LogAccess(req *http.Request, statusCode, byteSize int) {
+	if ac.commonAccessLogFilename != "" {
+		f, err := os.OpenFile(ac.commonAccessLogFilename, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+		if err != nil {
+			log.Warn("Can not open " + ac.commonAccessLogFilename + ": " + err.Error())
+			return
+		}
+		defer f.Close()
+		_, err = f.WriteString(ac.CommonLogFormat(req, statusCode, byteSize) + "\n")
+		if err != nil {
+			log.Warn("Can not write to " + ac.commonAccessLogFilename + ": " + err.Error())
+			return
+		}
+	}
+	if ac.combinedAccessLogFilename != "" {
+		f, err := os.OpenFile(ac.combinedAccessLogFilename, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
+		if err != nil {
+			log.Warn("Can not open " + ac.combinedAccessLogFilename + ": " + err.Error())
+			return
+		}
+		defer f.Close()
+		_, err = f.WriteString(ac.CombinedLogFormat(req, statusCode, byteSize) + "\n")
+		if err != nil {
+			log.Warn("Can not write to " + ac.combinedAccessLogFilename + ": " + err.Error())
+			return
+		}
+	}
+}